+++ tests/basic10.rs
@@ -1,0 +1,261 @@
+// Copyright 2016 Gomez Guillaume
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+mod common;
+use common::*;
+
+
+const SRC: &str = r#"// This file was generated by gir (https://github.com/gtk-rs/gir)
+impl Device {
+    /// Determines information about the current keyboard grab.
+    /// This is not public API and must not be used by applications.
+    ///
+    /// # Deprecated since 3.16
+    ///
+    /// The symbol was never meant to be used outside
+    ///  of GTK+
+    /// ## `display`
+    /// the display for which to get the grab information
+    /// ## `device`
+    /// device to get the grab information from
+    /// ## `grab_window`
+    /// location to store current grab window
+    /// ## `owner_events`
+    /// location to store boolean indicating whether
+    ///  the `owner_events` flag to `gdk_keyboard_grab` or
+    ///  `gdk_pointer_grab` was `true`.
+    ///
+    /// # Returns
+    ///
+    /// `true` if this application currently has the
+    ///  keyboard grabbed.
+    #[cfg_attr(feature = "v3_16", deprecated)]
+    pub fn grab_info_libgtk_only(display: &Display, device: &Device) -> Option<(Window, bool)> {
+        skip_assert_initialized!();
+        unsafe {
+            let mut grab_window = ptr::null_mut();
+            let mut owner_events = mem::uninitialized();
+            let ret = from_glib(gdk_sys::gdk_device_grab_info_libgtk_only(
+                display.to_glib_none().0,
+                device.to_glib_none().0,
+                &mut grab_window,
+                &mut owner_events,
+            ));
+            if ret {
+                Some((from_glib_none(grab_window), from_glib(owner_events)))
+            } else {
+                None
+            }
+        }
+    }
+    /// The ::changed signal is emitted either when the `Device`
+    /// has changed the number of either axes or keys. For example
+    /// In X this will normally happen when the slave device routing
+    /// events through the master device changes (for example, user
+    /// switches from the USB mouse to a tablet), in that case the
+    /// master device will change to reflect the new slave device
+    /// axes and keys.
+    pub fn connect_changed<F: Fn(&Device) + 'static>(&self, f: F) -> SignalHandlerId {
+        unsafe extern "C" fn changed_trampoline<F: Fn(&Device) + 'static>(
+            this: *mut gdk_sys::GdkDevice,
+            f: glib_sys::gpointer,
+        ) {
+            let f: &F = &*(f as *const F);
+            f(&from_glib_borrow(this))
+        }
+        unsafe {
+            let f: Box_<F> = Box_::new(f);
+            connect_raw(
+                self.as_ptr() as *mut _,
+                b"changed\0".as_ptr() as *const _,
+                Some(transmute(changed_trampoline::<F> as usize)),
+                Box_::into_raw(f),
+            )
+        }
+    }
+}
+"#;
+
+const SRC_STRIPPED: &str = r#"// This file was generated by gir (https://github.com/gtk-rs/gir)
+impl Device {
+    #[cfg_attr(feature = "v3_16", deprecated)]
+    pub fn grab_info_libgtk_only(display: &Display, device: &Device) -> Option<(Window, bool)> {
+        skip_assert_initialized!();
+        unsafe {
+            let mut grab_window = ptr::null_mut();
+            let mut owner_events = mem::uninitialized();
+            let ret = from_glib(gdk_sys::gdk_device_grab_info_libgtk_only(
+                display.to_glib_none().0,
+                device.to_glib_none().0,
+                &mut grab_window,
+                &mut owner_events,
+            ));
+            if ret {
+                Some((from_glib_none(grab_window), from_glib(owner_events)))
+            } else {
+                None
+            }
+        }
+    }
+    pub fn connect_changed<F: Fn(&Device) + 'static>(&self, f: F) -> SignalHandlerId {
+        unsafe extern "C" fn changed_trampoline<F: Fn(&Device) + 'static>(
+            this: *mut gdk_sys::GdkDevice,
+            f: glib_sys::gpointer,
+        ) {
+            let f: &F = &*(f as *const F);
+            f(&from_glib_borrow(this))
+        }
+        unsafe {
+            let f: Box_<F> = Box_::new(f);
+            connect_raw(
+                self.as_ptr() as *mut _,
+                b"changed\0".as_ptr() as *const _,
+                Some(transmute(changed_trampoline::<F> as usize)),
+                Box_::into_raw(f),
+            )
+        }
+    }
+}
+"#;
+
+fn get_md(file: &str) -> String {
+    let x = r###"
+<!-- impl Device::fn grab_info_libgtk_only -->
+Determines information about the current keyboard grab.
+This is not public API and must not be used by applications.
+
+# Deprecated since 3.16
+
+The symbol was never meant to be used outside
+ of GTK+
+## `display`
+the display for which to get the grab information
+## `device`
+device to get the grab information from
+## `grab_window`
+location to store current grab window
+## `owner_events`
+location to store boolean indicating whether
+ the `owner_events` flag to `gdk_keyboard_grab` or
+ `gdk_pointer_grab` was `true`.
+
+# Returns
+
+`true` if this application currently has the
+ keyboard grabbed.
+<!-- impl Device::fn connect_changed -->
+The ::changed signal is emitted either when the `Device`
+has changed the number of either axes or keys. For example
+In X this will normally happen when the slave device routing
+events through the master device changes (for example, user
+switches from the USB mouse to a tablet), in that case the
+master device will change to reflect the new slave device
+axes and keys.
+"###;
+    let mut y = format!("<!-- file {} -->", file);
+    y.push_str(x);
+    y
+}
+
+#[allow(unused_must_use)]
+#[test]
+fn strip() {
+    let test_file = "basic.rs";
+    let comment_file = "basic.md";
+    let temp_dir = tempdir().unwrap();
+    gen_file(&temp_dir, test_file, SRC);
+    {
+        let mut f = gen_file(&temp_dir, comment_file, "");
+        stripper_lib::strip_comments(temp_dir.path(), test_file, &mut f, false);
+    }
+    println!("Testing markdown");
+    compare_files(
+        &get_md(test_file),
+        &temp_dir.path().join(comment_file),
+    );
+    println!("Testing stripped file");
+    compare_files(SRC_STRIPPED, &temp_dir.path().join(test_file));
+}
+
+#[allow(unused_must_use)]
+#[test]
+fn regeneration() {
+    let test_file = "basic.rs";
+    let comment_file = "basic.md";
+    let temp_dir = tempdir().unwrap();
+    gen_file(&temp_dir, test_file, SRC_STRIPPED);
+    gen_file(&temp_dir, comment_file, &get_md(test_file));
+    stripper_lib::regenerate_doc_comments(
+        temp_dir.path().to_str().unwrap(),
+        false,
+        temp_dir.path().join(comment_file).to_str().unwrap(),
+        true,
+        false,
+    );
+    compare_files(SRC, &temp_dir.path().join(test_file));
+}
+
+#[allow(unused_must_use)]
+#[test]
+fn regeneration2() {
+    let test_file = "basic.rs";
+    let comment_file = "basic.md";
+    let temp_dir = tempdir().unwrap();
+    gen_file(&temp_dir, test_file, SRC_STRIPPED);
+    gen_file(&temp_dir, comment_file, &get_md(test_file));
+    stripper_lib::regenerate_doc_comments(
+        temp_dir.path().to_str().unwrap(),
+        false,
+        temp_dir.path().join(comment_file).to_str().unwrap(),
+        false,
+        true,
+    );
+    compare_files(SRC, &temp_dir.path().join(test_file));
+}
+
+#[allow(unused_must_use)]
+#[test]
+fn regeneration3() {
+    let test_file = "basic.rs";
+    let comment_file = "basic.md";
+    let temp_dir = tempdir().unwrap();
+    gen_file(&temp_dir, test_file, SRC_STRIPPED);
+    gen_file(&temp_dir, comment_file, &get_md(test_file));
+    stripper_lib::regenerate_doc_comments(
+        temp_dir.path().to_str().unwrap(),
+        false,
+        temp_dir.path().join(comment_file).to_str().unwrap(),
+        false,
+        false,
+    );
+    compare_files(SRC, &temp_dir.path().join(test_file));
+}
+
+#[allow(unused_must_use)]
+#[test]
+fn regeneration4() {
+    let test_file = "basic.rs";
+    let comment_file = "basic.md";
+    let temp_dir = tempdir().unwrap();
+    gen_file(&temp_dir, test_file, SRC_STRIPPED);
+    gen_file(&temp_dir, comment_file, &get_md(test_file));
+    stripper_lib::regenerate_doc_comments(
+        temp_dir.path().to_str().unwrap(),
+        false,
+        temp_dir.path().join(comment_file).to_str().unwrap(),
+        true,
+        true,
+    );
+    compare_files(SRC, &temp_dir.path().join(test_file));
+}
