+++ tests/basic11.rs
@@ -1,0 +1,184 @@
+// Copyright 2016 Gomez Guillaume
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+mod common;
+use common::*;
+
+const SRC: &str = r#"// This file was generated by gir (https://github.com/gtk-rs/gir)
+impl Device {
+    /// The ::changed signal is emitted either when the `Device`
+    /// has changed the number of either axes or keys. For example
+    /// In X this will normally happen when the slave device routing
+    /// events through the master device changes (for example, user
+    /// switches from the USB mouse to a tablet), in that case the
+    /// master device will change to reflect the new slave device
+    /// axes and keys.
+    pub fn connect_changed<F: Fn(&Device) + 'static>(&self, f: F) -> SignalHandlerId {
+        unsafe extern "C" fn changed_trampoline<F: Fn(&Device) + 'static>(
+            this: *mut gdk_sys::GdkDevice,
+            f: glib_sys::gpointer,
+        ) {
+            let f: &F = &*(f as *const F);
+            f(&from_glib_borrow(this))
+        }
+        unsafe {
+            let f: Box_<F> = Box_::new(f);
+            connect_raw(
+                self.as_ptr() as *mut _,
+                b"changed\0".as_ptr() as *const _,
+                Some(transmute(changed_trampoline::<F> as usize)),
+                Box_::into_raw(f),
+            )
+        }
+    }
+    /// The ::tool-changed signal is emitted on pen/eraser
+    /// ``GdkDevices`` whenever tools enter or leave proximity.
+    ///
+    /// Feature: `v3_22`
+    ///
+    /// ## `tool`
+    /// The new current tool
+    #[cfg(any(feature = "v3_22", feature = "dox"))]
+    pub fn connect_tool_changed<F: Fn(&Device, &DeviceTool) + 'static>(
+        &self,
+        f: F,
+    ) -> SignalHandlerId {
+        unsafe extern "C" fn tool_changed_trampoline<F: Fn(&Device, &DeviceTool) + 'static>(
+            this: *mut gdk_sys::GdkDevice,
+            tool: *mut gdk_sys::GdkDeviceTool,
+            f: glib_sys::gpointer,
+        ) {
+            let f: &F = &*(f as *const F);
+            f(&from_glib_borrow(this), &from_glib_borrow(tool))
+        }
+        unsafe {
+            let f: Box_<F> = Box_::new(f);
+            connect_raw(
+                self.as_ptr() as *mut _,
+                b"tool-changed\0".as_ptr() as *const _,
+                Some(transmute(tool_changed_trampoline::<F> as usize)),
+                Box_::into_raw(f),
+            )
+        }
+    }
+}
+"#;
+
+const SRC_STRIPPED: &str = r#"// This file was generated by gir (https://github.com/gtk-rs/gir)
+impl Device {
+    pub fn connect_changed<F: Fn(&Device) + 'static>(&self, f: F) -> SignalHandlerId {
+        unsafe extern "C" fn changed_trampoline<F: Fn(&Device) + 'static>(
+            this: *mut gdk_sys::GdkDevice,
+            f: glib_sys::gpointer,
+        ) {
+            let f: &F = &*(f as *const F);
+            f(&from_glib_borrow(this))
+        }
+        unsafe {
+            let f: Box_<F> = Box_::new(f);
+            connect_raw(
+                self.as_ptr() as *mut _,
+                b"changed\0".as_ptr() as *const _,
+                Some(transmute(changed_trampoline::<F> as usize)),
+                Box_::into_raw(f),
+            )
+        }
+    }
+    #[cfg(any(feature = "v3_22", feature = "dox"))]
+    pub fn connect_tool_changed<F: Fn(&Device, &DeviceTool) + 'static>(
+        &self,
+        f: F,
+    ) -> SignalHandlerId {
+        unsafe extern "C" fn tool_changed_trampoline<F: Fn(&Device, &DeviceTool) + 'static>(
+            this: *mut gdk_sys::GdkDevice,
+            tool: *mut gdk_sys::GdkDeviceTool,
+            f: glib_sys::gpointer,
+        ) {
+            let f: &F = &*(f as *const F);
+            f(&from_glib_borrow(this), &from_glib_borrow(tool))
+        }
+        unsafe {
+            let f: Box_<F> = Box_::new(f);
+            connect_raw(
+                self.as_ptr() as *mut _,
+                b"tool-changed\0".as_ptr() as *const _,
+                Some(transmute(tool_changed_trampoline::<F> as usize)),
+                Box_::into_raw(f),
+            )
+        }
+    }
+}
+"#;
+
+fn get_md(file: &str) -> String {
+    let x = r###"
+<!-- impl Device::fn connect_changed -->
+The ::changed signal is emitted either when the `Device`
+has changed the number of either axes or keys. For example
+In X this will normally happen when the slave device routing
+events through the master device changes (for example, user
+switches from the USB mouse to a tablet), in that case the
+master device will change to reflect the new slave device
+axes and keys.
+<!-- impl Device::fn connect_tool_changed -->
+The ::tool-changed signal is emitted on pen/eraser
+``GdkDevices`` whenever tools enter or leave proximity.
+
+Feature: `v3_22`
+
+## `tool`
+The new current tool
+"###;
+    let mut y = format!("<!-- file {} -->", file);
+    y.push_str(x);
+    y
+}
+
+#[allow(unused_must_use)]
+#[test]
+fn strip() {
+    let test_file = "basic.rs";
+    let comment_file = "basic.md";
+    let temp_dir = tempdir().unwrap();
+    gen_file(&temp_dir, test_file, SRC);
+    {
+        let mut f = gen_file(&temp_dir, comment_file, "");
+        stripper_lib::strip_comments(temp_dir.path(), test_file, &mut f, false);
+    }
+    println!("Testing markdown");
+    compare_files(
+        &get_md(test_file),
+        &temp_dir.path().join(comment_file),
+    );
+    println!("Testing stripped file");
+    compare_files(SRC_STRIPPED, &temp_dir.path().join(test_file));
+}
+
+#[allow(unused_must_use)]
+#[test]
+fn regeneration() {
+    let test_file = "basic.rs";
+    let comment_file = "basic.md";
+    let temp_dir = tempdir().unwrap();
+    gen_file(&temp_dir, test_file, SRC_STRIPPED);
+    gen_file(&temp_dir, comment_file, &get_md(test_file));
+    stripper_lib::regenerate_doc_comments(
+        temp_dir.path().to_str().unwrap(),
+        false,
+        temp_dir.path().join(comment_file).to_str().unwrap(),
+        true,
+        false,
+    );
+    compare_files(SRC, &temp_dir.path().join(test_file));
+}
